Struct dusk_jubjub::JubJubAffine

pub struct JubJubAffine { /* fields omitted */ }

This represents a Jubjub point in the affine (x, y) coordinates.

Implementations

impl JubJubAffine

pub const fn identity() -> Self

Constructs the neutral element (0, 1).

pub fn mul_by_cofactor(&self) -> JubJubExtended

Multiplies this point by the cofactor, producing an JubJubExtended

pub fn is_small_order(&self) -> Choice

Determines if this point is of small order.

pub fn is_torsion_free(&self) -> Choice

Determines if this point is torsion free and so is in the prime order subgroup.

pub fn is_prime_order(&self) -> Choice

Determines if this point is prime order, or in other words that the smallest scalar multiplied by this point that produces the identity is r. This is equivalent to checking that the point is both torsion free and not the identity.

pub fn get_x(&self) -> BlsScalar

Returns the x-coordinate of this point.

pub fn get_y(&self) -> BlsScalar

Returns the y-coordinate of this point.

pub const fn to_niels(&self) -> AffineNielsPoint

Performs a pre-processing step that produces an AffineNielsPoint for use in multiple additions.

pub const fn from_raw_unchecked(x: BlsScalar, y: BlsScalar) -> JubJubAffine

Constructs an JubJubAffine given x and y without checking that the point is on the curve.